requirements.txt
asgiref==3.11.1
Django==4.2.28
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.3.1
django-filter==24.1
drf-spectacular==0.27.0
psycopg2-binary==2.9.11
python-decouple==3.8
sqlparse==0.5.5
typing_extensions==4.15.0
gunicorn==21.2.0
python-dateutil==2.8.2
requests==2.31.0
Pillow==10.1.0
